Experimental blur filter based on FMF
The new filter now uses a FMF that manages to do less operations,
causing the number of memory accesses to drop from 271800 to 20340000
(2 orders of magnitude!).

A very crude attempt at measuring the time difference was also made
trough a benchmarking program, the time dropped from 640.8 ms ± 9.2 ms
to 599.7 ms ± 8.2 ms.

Not all edge cases are handled and there are still some optimizations
that can be experimented with.