Use texture clamping semantics when calculating large filters
Previously large filters (where the size of the window for the filter is
larger than the image itself) would simply be truncated to the image
dimensions.

This works for images that assume a static border outside of the image
bounds, but for images that assume border clamping semantics (ie. the
value of the nearest pixel is read) this doesn't work because the avg
signal of the extrapolated pixels isn't constant for any extensions (for
example corners have a larger influence).

To fix this the extrapolated pixels are simulated by multypling the last
pixel in bounds by 1 + the number of OOB pixels.